feat(Q2): resolving negative yiq but with smurf mounth and refacting some convert function code
