+++ Q2/helper.py
@@ -1,0 +1,20 @@
+import numpy as np
+
+def transformRGB2YIQ(imgRGB: np.ndarray) -> np.ndarray:
+    yiq_from_rgb = np.array([[0.299, 0.587, 0.114],
+                             [0.59590059, -0.27455667, -0.32134392],
+                             [0.21153661, -0.52273617, 0.31119955]])
+    OrigShape=imgRGB.shape
+    return np.dot(imgRGB.reshape(-1,3), yiq_from_rgb.transpose()).reshape(OrigShape)
+
+    pass
+
+
+def transformYIQ2RGB(imgYIQ: np.ndarray) -> np.ndarray:
+    yiq_from_rgb = np.array([[0.299, 0.587, 0.114],
+                             [0.59590059, -0.27455667, -0.32134392],
+                             [0.21153661, -0.52273617, 0.31119955]])
+    OrigShape=imgYIQ.shape
+    return np.dot(imgYIQ.reshape(-1,3), np.linalg.inv(yiq_from_rgb).transpose()).reshape(OrigShape)
+
+    pass
